got fundraising projections to work at ghci
20160723-19:02:03 mengwong@venice2:~/non-db-src/l/compiler/sandbox5-rounds/fundraising.hsproj% cd .; ghci Fundraising.hs
GHCi, version 7.10.3: http://www.haskell.org/ghc/  :? for help
[1 of 1] Compiling Main             ( Fundraising.hs, interpreted )
Ok, modules loaded: Main.
*Main> project ValGrowth { years=10, startValuation =   4000000, endValuation   = 5000000000 }
If initial valuation is 4,000,000 and want to get to 5,000,000,000 in 10 years,
then every year our valuation will need to increase by 2.0402858 times.

in 2016, we will be worth      4,000,000
in 2017, we will be worth      8,161,143
in 2018, we will be worth     16,651,066
in 2019, we will be worth     33,972,932
in 2020, we will be worth     69,314,488
in 2021, we will be worth    141,421,376
in 2022, we will be worth    288,540,032
in 2023, we will be worth    588,704,128
in 2024, we will be worth  1,201,124,608
in 2025, we will be worth  2,450,637,568
in 2026, we will be worth  5,000,001,536

*Main>
Leaving GHCi.

+++ sandbox5-rounds/fundraising.hsproj/Fundraising.hs
@@ -125,7 +125,7 @@ project valgrowth = do
   putStrLn $ show valgrowth
   putStrLn $ unlines $
       map (\a -> printf "in %d, we will be worth %14s"
-           (2016+a)
+           ((2016+a)::Int)
            (digify $ truncate (fromIntegral
             (startValuation valgrowth) *
              (yearlyGrowth valgrowth)
